refactor(oid4vc): split public_routes.py and routes.py into focused submodules
Split two large monolithic files into organized package directories:

public_routes.py (953 lines) -> public_routes/ package:
  - credential_offer.py: credential offer dereference endpoint
  - metadata.py: credential issuer metadata endpoint
  - nonce.py: nonce creation and request
  - notification.py: notification endpoint
  - token.py: token endpoint, check_token, proof of possession
  - credential.py: credential issuance endpoint
  - verification.py: OID4VP verification (get_request, post_response)
  - status_list.py: status list endpoint

routes.py (1854 lines) -> routes/ package:
  - exchange.py: exchange record CRUD
  - credential_offer.py: credential offer helpers and endpoints
  - supported_credential.py: supported credential CRUD
  - vp_request.py: VP request creation and listing
  - vp_dcql.py: DCQL query CRUD
  - vp_pres_def.py: presentation definition CRUD
  - vp_presentation.py: VP presentation listing and retrieval
  - did_jwk.py: DID JWK creation
  - helpers.py: issuer config endpoints

Both __init__.py files re-export all public symbols so existing
imports (from oid4vc.routes import X, from oid4vc.public_routes import Y)
continue to work without changes.
